Warn and stop buckets when overwritten
When adding a bucket with an ID that's already used, currently the
previous bucket is just silently overwritten. This causes issues if the
bucket has for example an interval trigger, since that continues to run
unchecked indefinitely, which was the original issue described in #13
This change stops any overwritten bucket and logs a warning.